rename 'readers'->'setup'
